# python-flask-cache/swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Swagger Petstore
  description: "This is a sample Petstore server.  You can find\nout more about Swagger\
    \ at\n[http://swagger.io](http://swagger.io) or on\n[irc.freenode.net, #swagger](http://swagger.io/irc/).\n"
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
paths:
...
  /api/users:
    get:
      tags:
      - users
      summary: the method to show all users
      description: you can show all users
      operationId: get_all_users
      parameters:
      - name: key_word
        in: query
        description: you can fill key word you want to search
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: page_num
        in: query
        description: number of page
        required: false
        style: form
        explode: true
        schema:
          type: number
      - name: records_per_page
        in: query
        description: number record in a page
        required: false
        style: form
        explode: true
        schema:
          type: number
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/users'
                x-content-type: application/json
        "404":
          description: not found
      security:
      - bearer_auth: []
      x-openapi-router-controller: swagger_server.controllers.users_controller
    put:
      tags:
      - users
      summary: the method to register
      description: you can register user
      operationId: update_user
      requestBody:
        description: body
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/users'
        required: true
      responses:
        "200":
          description: created success
        "400":
          description: bad request
      security:
      - bearer_auth: []
      x-openapi-router-controller: swagger_server.controllers.users_controller
    post:
      tags:
      - users
      summary: the method to register
      description: you can register user
      operationId: create_user
      requestBody:
        description: body
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/users'
        required: true
      responses:
        "200":
          description: created success
        "400":
          description: bad request
      security:
      - bearer_auth: []
      x-openapi-router-controller: swagger_server.controllers.users_controller
  /api/users/{user_id}:
    get:
      tags:
      - users
      summary: Show detail a user by id
      description: method to show info a user by student_id
      operationId: get_user_by_id
      parameters:
      - name: user_id
        in: path
        description: fill in id a student
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: infor a student
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/users'
            application/xml:
              schema:
                $ref: '#/components/schemas/users'
        "400":
          description: bad request
      security:
      - bearer_auth: []
      x-openapi-router-controller: swagger_server.controllers.users_controller
    delete:
      tags:
      - users
      summary: delete a user by id
      description: method to delete a user by user_id
      operationId: del_user_by_id
      parameters:
      - name: user_id
        in: path
        description: fill in id a user
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: success
        "400":
          description: bad request
      security:
      - bearer_auth: []
      x-openapi-router-controller: swagger_server.controllers.users_controller
  /api/user/login:
    post:
      tags:
      - users
      summary: log user into your system
      description: method to login user
      operationId: login_user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/body'
      responses:
        "200":
          description: login success
          headers:
            Set-Cookie:
              style: simple
              explode: false
              schema:
                type: string
                example: JSESSIONID = abcd12345 ; Path=/; HttpOnly
        "403":
          description: csrf_token fobbiden
      x-openapi-router-controller: swagger_server.controllers.users_controller
...
components:
  schemas:
...
    users:
      required:
      - email
      - fullname
      - password
      - username
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        fullname:
          type: string
        birthday:
          type: string
          format: date
        last_login_date:
          type: string
          format: date-time
        lockout_date:
          type: string
          format: date-time
        login_failed_count:
          type: number
        register_date:
          type: string
          format: date-time
        forgot_password_token:
          type: string
          format: base64
      description: the table user
      example:
        birthday: 2000-01-23
        password: ""
        login_failed_count: 6.02745618307040320615897144307382404804229736328125
        lockout_date: 2000-01-23T04:56:07.000+00:00
        register_date: 2000-01-23T04:56:07.000+00:00
        id: 0
        fullname: fullname
        forgot_password_token: forgot_password_token
        last_login_date: 2000-01-23T04:56:07.000+00:00
        email: ""
        username: username
      xml:
        name: users
    body:
      type: object
      properties:
        username:
          type: string
          description: fill username
        password:
          type: string
          description: fill password
          format: password
...
  securitySchemes:
    bearer_auth:
      type: http
      description: authentication bearer
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: swagger_server.controllers.authorization_controller.check_bearer_auth